Continue with updating LookupTable
Finished adding tests of the model.  I'm not thrilled with the
ExpressionParser class being used directly by LookupTable, but changing
it to use DI is beyond the scope of this issue.  When the LookupTable
changes its implementation to a sorted array, the dependency can be
changed (perhaps).

Should the MTscript functions have tests as well?  Or should those be
done via a macro within the tool?  It would certainly be easier to
automate if they were in Java, but putting a bunch of testing macros
into a campaign would (potentially) allow others to build tests as well
and contribute to the project.

(TODO) With the use of `macro.catch`, maybe exceptions thrown within
MTscript can actually be caught and handled in a sufficient way?